+++ Core/Components/Agnus/DmaDebugger/DmaDebugger.cpp
@@ -368,7 +368,7 @@ DmaDebugger::hsyncHandler(isize vpos)
 void
 DmaDebugger::computeOverlay(Texel *ptr, isize first, isize last, BusOwner *own, u16 *val)
 {
-    double opacity = config.opacity / 100.0;
+    double opacity = double(config.opacity) / 100.0;
     double bgWeight = 0;
     double fgWeight = 0;
